+++ Service/Sources/Data/DataSource/Auth/LocalAuthDataSourceImpl.swift
@@ -1,0 +1,24 @@
+import Foundation
+
+struct LocalAuthDataSourceImpl: LocalAuthDataSource {
+    private let keychain: any Keychain
+
+    public init(keychain: any Keychain) {
+        self.keychain = keychain
+    }
+
+    public func logout() {
+        keychain.delete(type: .accessToken)
+        keychain.delete(type: .accessExpiredAt)
+        keychain.delete(type: .refreshToken)
+        keychain.delete(type: .refreshExpiredAt)
+    }
+
+    public func withdraw() {
+        keychain.delete(type: .accessToken)
+        keychain.delete(type: .accessExpiredAt)
+        keychain.delete(type: .refreshToken)
+        keychain.delete(type: .refreshExpiredAt)
+    }
+}
+
